requirements.txt
discord.py~=2.5.2
requests~=2.32.3
flask~=3.1.0
language_tool_python~=2.9.4
wikipedia-api~=0.7.1
protobuf~=4.25.5
aiohttp~=3.10.11
mistralai~=1.9.1
beautifulsoup4~=4.13.4
groq~=0.30.0